Extract trace writer queue helpers
